refactor: Move element parent tests to ParentTest
